+++ tel-testdata/tel-testdata-integrations/src/main/java/dk/sunepoulsen/tel/testdata/module/integrator/model/DataPoint.java
@@ -9,11 +9,19 @@ import lombok.Data;
 @Data
 @Schema(name = "Data Point", description = "Data point of x and y values")
 public class DataPoint implements BaseModel {
+    @Schema(
+        description = "X-coordinate of a data point",
+        requiredMode = Schema.RequiredMode.REQUIRED,
+        accessMode = Schema.AccessMode.READ_WRITE
+    )
     @NotNull
-    @Parameter(description = "X coordinate", required = true)
     private Double x;
 
+    @Schema(
+        description = "Y-coordinate of a data point",
+        requiredMode = Schema.RequiredMode.REQUIRED,
+        accessMode = Schema.AccessMode.READ_WRITE
+    )
     @NotNull
-    @Parameter(description = "Y coordinate", required = true)
     private Double y;
 }