+++ tel-testdata/tel-testdata-integrations/src/main/java/dk/sunepoulsen/tel/testdata/module/integrator/model/DataSet.java
@@ -4,7 +4,6 @@ import dk.sunepoulsen.tes.rest.models.BaseModel;
 import dk.sunepoulsen.tes.rest.models.validation.annotations.OnCrudCreate;
 import dk.sunepoulsen.tes.rest.models.validation.annotations.OnCrudRead;
 import dk.sunepoulsen.tes.rest.models.validation.annotations.OnCrudUpdate;
-import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Null;
@@ -13,16 +12,28 @@ import lombok.Data;
 @Data
 @Schema(name = "Dataset", description = "Defines a data set")
 public class DataSet implements BaseModel {
+    @Schema(
+        description = "Unique id of the data set",
+        requiredMode = Schema.RequiredMode.REQUIRED,
+        accessMode = Schema.AccessMode.READ_ONLY
+    )
     @Null(groups = {OnCrudCreate.class, OnCrudUpdate.class})
     @NotNull(groups = {OnCrudRead.class})
-    @Parameter(description = "Unique id of the data set", required = true)
     private Long id;
 
+    @Schema(
+        description = "Name of the data set",
+        requiredMode = Schema.RequiredMode.REQUIRED,
+        accessMode = Schema.AccessMode.READ_WRITE
+    )
     @NotNull(groups = {OnCrudCreate.class, OnCrudRead.class})
-    @Parameter(description = "Name of the data set", required = true)
     private String name;
 
+    @Schema(
+        description = "Description of the data set",
+        requiredMode = Schema.RequiredMode.REQUIRED,
+        accessMode = Schema.AccessMode.READ_WRITE
+    )
     @NotNull(groups = {OnCrudCreate.class, OnCrudRead.class})
-    @Parameter(description = "Description of the data set", required = true)
     private String description;
 }