+++ tel-testdata/CHANGELOG.md
@@ -19,6 +19,7 @@ For release-dates, use date-format: YYYY-MM-DD
 
 - [TESENTLABS-2](https://sunepoulsen.atlassian.net/browse/TESENTLABS-2): Create testdata backend service
 - [TESENTLABS-27](https://sunepoulsen.atlassian.net/browse/TESENTLABS-27): Manage data sets in tel-testdata
+  - Add api model for the endpoints `POST /datasets/data-points`
 
 ### Fixed